improve dependency tree and file edits
