+++ packages/cli/package.json
@@ -75,7 +75,6 @@
   },
   "dependencies": {
     "axios": "^1.7.7",
-    "dependency-tree": "^11.0.1",
     "express": "^4.21.1",
     "http-proxy-middleware": "^3.0.3",
     "openai": "^4.67.2",
