+++ packages/cli/src/api/scan.ts
@@ -1,12 +1,11 @@
 import { z } from "zod";
-import { getDependencyTree } from "../helper/dependencies";
-import { ScanCodebaseResponsePayload } from "../helper/payloads";
-import { getEndpontsFromTree } from "../helper/tree";
+import {
+  getDependencyTree,
+  getEndpontsFromTree,
+} from "../helper/dependencyTree";
 import { scanSchema } from "./helpers/validation";
 
-export function scan(
-  payload: z.infer<typeof scanSchema>,
-): ScanCodebaseResponsePayload {
+export function scan(payload: z.infer<typeof scanSchema>) {
   const tree = getDependencyTree(payload.entrypointPath);
 
   const endpoints = getEndpontsFromTree(tree);
